+++ src/Composer/Installers/LavaLiteInstaller.php
@@ -1,0 +1,10 @@
+<?php
+namespace Composer\Installers;
+
+class LavaLiteInstaller extends BaseInstaller
+{
+    protected $locations = array(
+        'package' => 'packages/{$name}/',
+        'theme'   => 'public/themes/{$name}/',
+    );
+}